refactor: MovingPosition을 활용해서 Service, DAO 로직을 수행하도록 수정
ChessWebService MovingPosition을 활용해서 로직을 수행
HistoryDao MovingPosition을 활용해서 로직을 수행
ALL 패키지 이동에 따른 리팩토링
